+++ xslang/sources/hir/type_check/expression_type.rs
@@ -1,0 +1,29 @@
+/*
+ * SPDX-FileCopyrightText: 2026 Leitwolf <xs-lang.chess031@slmails.com>
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+use super::*;
+
+impl TypeChecker
+{
+  pub(super) fn expression_type(&self, expression: &Expression) -> Option<Type>
+  {
+    match expression
+    {
+      Expression::Literal { literal, .. } => literal_default_type(literal),
+      Expression::Local { name, .. } => self.find_local(name).map(|local| local.ty.clone()),
+      Expression::Assign { value, .. } => self.expression_type(value),
+      Expression::Binary { operator,
+                           left,
+                           right,
+                           .. } => self.binary_expression_type(*operator, left, right),
+      Expression::ResultPropagation { value, .. } => self.result_success_type(value),
+      Expression::Call { return_type, .. } |
+      Expression::If { result_type: return_type,
+                       .. } |
+      Expression::Match { result_type: return_type,
+                          .. } => Some(return_type.as_ref().clone()),
+    }
+  }
+}